+++ govc/test/test_helper.bash
@@ -101,6 +101,19 @@ vcsim_env() {
   fi
 }
 
+# remove username/password from $GOVC_URL and set $GOVC_{USERNAME,PASSWORD}
+govc_url_to_vars() {
+  local url=$(awk -F@ '{print $2}' <<<"$GOVC_URL")
+  local userpass=$(awk -F// '{print $2}' <<<"$GOVC_URL" | awk -F@ '{print $1}')
+  local username=$(awk -F:  '{print $1}' <<<"$userpass")
+  local password=$(awk -F:  '{print $2}' <<<"$userpass")
+
+  export GOVC_URL="$url" GOVC_USERNAME="$username" GOVC_PASSWORD="$password"
+
+  # double check that we removed user/pass
+  grep -q -v @ <<<"$GOVC_URL"
+}
+
 quit_vnc() {
   if [ "$(uname)" = "Darwin" ]; then
     osascript <<EOF