+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### (unreleased)
 
+* Add LoginExtensionByCertificate support
+
+* Add object.ExtensionManager
+
 ...
 
 ### 0.2.0 (2015-09-15)